sshdgenkeys.service: do not use RemainAfterExit=
RemainAfterExit=yes is pretty pointless here, but rather
causes some inconsistency: if the conditions don't pass,
the service stays inactive, which means when sshd.service
is restarted and new key types got added, the service
would actually be triggered. OTOH, if the conditions
don't pass initially, the service stays active after
doing its job, and sshd.service restarts would not result
in reevaluation of the sshdgenkeys.service conditions.

Therefore, the behavior is mostly not useful, and sometimes
could lead to confusion. Drop it.